+++ admin/footer.php
@@ -2,5 +2,7 @@
 			<div class="copyright"><?php RSCopyright(); ?></div>
 			<div class="version"><?php RSVersion(); ?></div>
 		</footer>
+		<?php getScript('jquery.min.js', '3.4.1'); ?>
+		<?php getAdminScript('script.js'); ?>
 	</body>
 </html>